try to save space in page

+++ srcs/page_info.c
@@ -6,30 +6,30 @@
 /*   By: amoinier <amoinier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/22 17:17:21 by amoinier          #+#    #+#             */
-/*   Updated: 2017/11/22 17:18:58 by amoinier         ###   ########.fr       */
+/*   Updated: 2017/11/23 14:28:55 by amoinier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "malloc.h"
 
-size_t 	get_page_size(size_t size)
+unsigned int 	get_page_size(size_t size)
 {
 	if (size <= TINY_SIZE)
-		return (TINY);
+		return ((int)TINY);
 	else if (size <= SMALL_SIZE)
-		return (SMALL);
+		return ((int)SMALL);
 	else
-		return (size);
+		return ((int)size);
 }
 
-size_t 	get_block_size(size_t size)
+unsigned int 	get_block_size(size_t size)
 {
 	if (size <= TINY_SIZE)
-		return (TINY_SIZE);
+		return ((int)TINY_SIZE);
 	else if (size <= SMALL_SIZE)
-		return (SMALL_SIZE);
+		return ((int)SMALL_SIZE);
 	else
-		return (size);
+		return ((int)size);
 }
 
 int		get_page_type(size_t size)
